move var_config.Variable to modconfig and rename RawVariable move variable parsing to parse
