+++ hclhelpers/expression.go
@@ -1,18 +1,18 @@
-package var_config
+package hclhelpers
 
 import (
 	"github.com/hashicorp/hcl/v2"
 	"github.com/hashicorp/hcl/v2/hclsyntax"
 )
 
-// exprIsNativeQuotedString determines whether the given expression looks like
+// ExprIsNativeQuotedString determines whether the given expression looks like
 // it's a quoted string in the HCL native syntax.
 //
 // This should be used sparingly only for situations where our legacy HCL
 // decoding would've expected a keyword or reference in quotes but our new
 // decoding expects the keyword or reference to be provided directly as
 // an identifier-based expression.
-func exprIsNativeQuotedString(expr hcl.Expression) bool {
+func ExprIsNativeQuotedString(expr hcl.Expression) bool {
 	_, ok := expr.(*hclsyntax.TemplateExpr)
 	return ok
 }
