--- modconfig/var_config/doc.go
@@ -1,3 +1,0 @@
-// Package variable_config contains code from https://github.com/hashicorp/terraform, with minor modifications
-
-package var_config
