+++ parse/variable.go
@@ -1,59 +1,34 @@
-package var_config
+package parse
 
-// github.com/turbot/terraform-components/configs/parser_config.go
 import (
 	"fmt"
-	"unicode"
-
-	"github.com/turbot/pipe-fittings/hclhelpers"
-	"github.com/turbot/pipe-fittings/schema"
-
 	"github.com/hashicorp/hcl/v2"
 	"github.com/hashicorp/hcl/v2/ext/typeexpr"
 	"github.com/hashicorp/hcl/v2/gohcl"
 	"github.com/hashicorp/hcl/v2/hclsyntax"
+	"github.com/turbot/pipe-fittings/hclhelpers"
+	"github.com/turbot/pipe-fittings/modconfig"
+	"github.com/turbot/pipe-fittings/schema"
 	"github.com/zclconf/go-cty/cty"
 	"github.com/zclconf/go-cty/cty/convert"
 )
 
-// A consistent detail message for all "not a valid identifier" diagnostics.
-const badIdentifierDetail = "A name must start with a letter or underscore and may contain only letters, digits, underscores, and dashes."
-
-// Variable represents a "variable" block in a module or file.
-type Variable struct {
-	Name           string
-	Title          string
-	Description    string
-	Default        cty.Value
-	Type           cty.Type
-	ParsingMode    VariableParsingMode
-	Enum           cty.Value
-	EnumGo         []any
-	DescriptionSet bool
-	DeclRange      hcl.Range
-}
-
-func DecodeVariableBlock(block *hcl.Block, content *hcl.BodyContent, override bool) (*Variable, hcl.Diagnostics) {
-	v := &Variable{
+func DecodeVariableBlock(block *hcl.Block, content *hcl.BodyContent) (*modconfig.RawVariable, hcl.Diagnostics) {
+	v := &modconfig.RawVariable{
 		Name:      block.Labels[0],
 		DeclRange: hclhelpers.BlockRange(block),
 	}
 	var diags hcl.Diagnostics
 
-	// Unless we're building an override, we'll set some defaults
-	// which we might override with attributes below. We leave these
-	// as zero-value in the override case so we can recognize whether
-	// or not they are set when we merge.
-	if !override {
-		v.Type = cty.DynamicPseudoType
-		v.ParsingMode = VariableParseLiteral
-	}
+	//  set some defaults which we might override with attributes below.
+	v.Type = cty.DynamicPseudoType
+	v.ParsingMode = modconfig.VariableParseLiteral
 
 	if !hclsyntax.ValidIdentifier(v.Name) {
 		diags = append(diags, &hcl.Diagnostic{
 			Severity: hcl.DiagError,
 			Summary:  "Invalid variable name",
-			Detail:   badIdentifierDetail,
+			Detail:   modconfig.BadIdentifierDetail,
 			Subject:  &block.LabelRanges[0],
 		})
 	}
@@ -208,11 +183,11 @@ func DecodeVariableBlock(block *hcl.Block, content *hcl.BodyContent, override bo
 	return v, diags
 }
 
-func decodeVariableType(expr hcl.Expression) (cty.Type, VariableParsingMode, hcl.Diagnostics) {
-	if exprIsNativeQuotedString(expr) {
+func decodeVariableType(expr hcl.Expression) (cty.Type, modconfig.VariableParsingMode, hcl.Diagnostics) {
+	if hclhelpers.ExprIsNativeQuotedString(expr) {
 		val, diags := expr.Value(nil)
 		if diags.HasErrors() {
-			return cty.DynamicPseudoType, VariableParseHCL, diags
+			return cty.DynamicPseudoType, modconfig.VariableParseHCL, diags
 		}
 		str := val.AsString()
 		switch str {
@@ -222,23 +197,23 @@ func decodeVariableType(expr hcl.Expression) (cty.Type, VariableParsingMode, hcl
 				Summary:  "Invalid quoted type constraints",
 				Subject:  expr.Range().Ptr(),
 			})
-			return cty.DynamicPseudoType, VariableParseLiteral, diags
+			return cty.DynamicPseudoType, modconfig.VariableParseLiteral, diags
 		case "list":
 			diags = append(diags, &hcl.Diagnostic{
 				Severity: hcl.DiagError,
 				Summary:  "Invalid quoted type constraints",
 				Subject:  expr.Range().Ptr(),
 			})
-			return cty.DynamicPseudoType, VariableParseHCL, diags
+			return cty.DynamicPseudoType, modconfig.VariableParseHCL, diags
 		case "map":
 			diags = append(diags, &hcl.Diagnostic{
 				Severity: hcl.DiagError,
 				Summary:  "Invalid quoted type constraints",
 				Subject:  expr.Range().Ptr(),
 			})
-			return cty.DynamicPseudoType, VariableParseHCL, diags
+			return cty.DynamicPseudoType, modconfig.VariableParseHCL, diags
 		default:
-			return cty.DynamicPseudoType, VariableParseHCL, hcl.Diagnostics{{
+			return cty.DynamicPseudoType, modconfig.VariableParseHCL, hcl.Diagnostics{{
 				Severity: hcl.DiagError,
 				Summary:  "Invalid legacy variable type hint",
 				Subject:  expr.Range().Ptr(),
@@ -252,120 +227,22 @@ func decodeVariableType(expr hcl.Expression) (cty.Type, VariableParsingMode, hcl
 	// elements are consistent. This is the same as list(any) or map(any).
 	switch hcl.ExprAsKeyword(expr) {
 	case "list":
-		return cty.List(cty.DynamicPseudoType), VariableParseHCL, nil
+		return cty.List(cty.DynamicPseudoType), modconfig.VariableParseHCL, nil
 	case "map":
-		return cty.Map(cty.DynamicPseudoType), VariableParseHCL, nil
+		return cty.Map(cty.DynamicPseudoType), modconfig.VariableParseHCL, nil
 	}
 
 	ty, diags := typeexpr.TypeConstraint(expr)
 	if diags.HasErrors() {
-		return cty.DynamicPseudoType, VariableParseHCL, diags
+		return cty.DynamicPseudoType, modconfig.VariableParseHCL, diags
 	}
 
 	switch {
 	case ty.IsPrimitiveType():
 		// Primitive types use literal parsing.
-		return ty, VariableParseLiteral, diags
+		return ty, modconfig.VariableParseLiteral, diags
 	default:
 		// Everything else uses HCL parsing
-		return ty, VariableParseHCL, diags
+		return ty, modconfig.VariableParseHCL, diags
 	}
-}
-
-// Required returns true if this variable is required to be set by the caller,
-// or false if there is a default value that will be used when it isn't set.
-func (v *Variable) Required() bool {
-	return v.Default == cty.NilVal
-}
-
-// VariableParsingMode defines how values of a particular variable given by
-// text-only mechanisms (command line arguments and environment variables)
-// should be parsed to produce the final value.
-type VariableParsingMode rune
-
-// VariableParseLiteral is a variable parsing mode that just takes the given
-// string directly as a cty.String value.
-const VariableParseLiteral VariableParsingMode = 'L'
-
-// VariableParseHCL is a variable parsing mode that attempts to parse the given
-// string as an HCL expression and returns the result.
-const VariableParseHCL VariableParsingMode = 'H'
-
-// Parse uses the receiving parsing mode to process the given variable value
-// string, returning the result along with any diagnostics.
-//
-// A VariableParsingMode does not know the expected type of the corresponding
-// variable, so it's the caller's responsibility to attempt to convert the
-// result to the appropriate type and return to the user any diagnostics that
-// conversion may produce.
-//
-// The given name is used to create a synthetic filename in case any diagnostics
-// must be generated about the given string value. This should be the name
-// of the configuration variable whose value will be populated from the given
-// string.
-//
-// If the returned diagnostics has errors, the returned value may not be
-// valid.
-func (m VariableParsingMode) Parse(name, value string) (cty.Value, hcl.Diagnostics) {
-	switch m {
-	case VariableParseLiteral:
-		return cty.StringVal(value), nil
-	case VariableParseHCL:
-		fakeFilename := fmt.Sprintf("<value for var.%s>", name)
-		expr, diags := hclsyntax.ParseExpression([]byte(value), fakeFilename, hcl.Pos{Line: 1, Column: 1})
-		if diags.HasErrors() {
-			return cty.DynamicVal, diags
-		}
-		val, valDiags := expr.Value(nil)
-		diags = append(diags, valDiags...)
-		return val, diags
-	default:
-		// Should never happen
-		panic(fmt.Errorf("parse called on invalid VariableParsingMode %#v", m))
-	}
-}
-
-// VariableValidation represents a configuration-defined validation rule
-// for a particular input variable, given as a "validation" block inside
-// a "variable" block.
-type VariableValidation struct {
-	// Condition is an expression that refers to the variable being tested
-	// and contains no other references. The expression must return true
-	// to indicate that the value is valid or false to indicate that it is
-	// invalid. If the expression produces an error, that's considered a bug
-	// in the module defining the validation rule, not an error in the caller.
-	Condition hcl.Expression
-
-	// ErrorMessage is one or more full sentences, which would need to be in
-	// English for consistency with the rest of the error message output but
-	// can in practice be in any language as long as it ends with a period.
-	// The message should describe what is required for the condition to return
-	// true in a way that would make sense to a caller of the module.
-	ErrorMessage string
-
-	DeclRange hcl.Range
-}
-
-// looksLikeSentence is a simple heuristic that encourages writing error
-// messages that will be presentable when included as part of a larger error diagnostic
-func looksLikeSentences(s string) bool {
-	if len(s) < 1 {
-		return false
-	}
-	runes := []rune(s) // HCL guarantees that all strings are valid UTF-8
-	first := runes[0]
-	last := runes[len(runes)-1]
-
-	// If the first rune is a letter then it must be an uppercase letter.
-	// (This will only see the first rune in a multi-rune combining sequence,
-	// but the first rune is generally the letter if any are, and if not then
-	// we'll just ignore it because we're primarily expecting English messages
-	// right now anyway)
-	if unicode.IsLetter(first) && !unicode.IsUpper(first) {
-		return false
-	}
-
-	// The string must be at least one full sentence, which implies having
-	// sentence-ending punctuation.
-	return last == '.' || last == '?' || last == '!'
 }
